フィールド変数の修正

+++ src/main/java/com/example/demo/domain/grade/model/propagation/UserGradeClassView.java
@@ -4,7 +4,7 @@ public interface UserGradeClassView {
 
 	Integer getYear();
 
-	Byte getGrade();
+	Integer getGrade();
 
 	String getClazzName();
 }
